requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-filter==23.2
django-money==3.2.0
djangorestframework==3.14.0
Faker==19.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.27.8
Pillow==10.0.0
py-moneyed==3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
